# Conflicts:
#   src/main/java/com/gameplat/admin/controller/open/chat/ChatNoticeController.java

+++ src/main/java/com/gameplat/admin/model/dto/ActivityLobbyDiscountDTO.java
@@ -24,14 +24,14 @@ public class ActivityLobbyDiscountDTO implements Serializable {
   private String discountUrl;
 
   @Schema(description = "目标值")
-  private Integer targetValue;
+  private Long targetValue;
 
   @Schema(description = "赠送值")
-  private Integer presenterValue;
+  private BigDecimal presenterValue;
 
   @Schema(description = "赠送打码")
   private BigDecimal presenterDml;
 
   @Schema(description = "提现打码")
-  private Integer withdrawDml;
+  private BigDecimal withdrawDml;
 }
